refactor: update file structure and imports (#91)
